requirements.txt
aiohttp
aiohttp_cors
aiohttp-jinja2

onnxruntime
numpy
pillow
